+++ src/nodejs/lambda-layers/database-util/src/query/submission.js
@@ -4,13 +4,14 @@ const workflow = require('./workflow.js');
 
 const table = 'submission';
 // const allFields = ['id', 'name', 'user_id', 'daac_id', 'conversation_id', 'workflow_id', 'workflow_name', 'step_name', 'status', 'forms', 'action_data', 'form_data', 'metadata', 'created_at', 'last_change', 'lock'];
-const allFields = ['id', 'name', 'initiator', 'workflow_id', 'hidden', 'conversation_id', 'workflow_name', 'daac_id', 'step_data', 'step_name', 'status', 'forms', 'action_data', 'form_data', 'metadata', 'created_at', 'last_change', 'lock', 'contributor_ids', 'copy', 'origin_id'];
+const allFields = ['id', 'name', 'initiator', 'workflow_id', 'hidden', 'conversation_id', 'workflow_name', 'daac_id', 'daac_name', 'step_data', 'step_name', 'status', 'forms', 'action_data', 'form_data', 'metadata', 'created_at', 'last_change', 'lock', 'contributor_ids', 'copy', 'origin_id'];
 const fieldMap = {
   id: 'submission.id',
   name: 'submission.name',
   initiator: 'initiator_ref.initiator',
   user_id: 'submission.initiator_edpuser_id user_id',
   daac_id: 'submission.daac_id',
+  daac_name: 'daac.long_name daac_name',
   conversation_id: 'submission.conversation_id',
   hidden: 'submission.hidden',
   workflow_id: 'submission_status.workflow_id',
@@ -152,6 +153,11 @@ const refs = {
     src: 'workflow',
     on: { left: 'workflow.id', right: 'submission_status.workflow_id' }
   },
+  daac: {
+    type: 'left_join',
+    src: 'daac',
+    on: { left: 'daac.id', right: 'submission.daac_id' }
+  },
   submission_form_data_pool: {
     type: 'left_join',
     src: 'submission_form_data_pool',
@@ -165,7 +171,7 @@ const findById = (params) => sql.select({
   fields: fields(allFields),
   from: {
     base: table,
-    joins: [refs.submission_status, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, refs.submission_form_data, refs.step, refs.workflow, refs.submission_form_data_pool, refs.submission_copy]
+    joins: [refs.submission_status, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, refs.submission_form_data, refs.step, refs.workflow, refs.daac, refs.submission_form_data_pool, refs.submission_copy]
   },
   where: {
     filters: [{ field: fieldMap.id, param: 'id' }]
@@ -176,7 +182,7 @@ const getUsersSubmissions = (params) => sql.select({
   fields: fields(allFields),
   from: {
     base: table,
-    joins: [refs.submission_status, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, refs.submission_form_data, refs.step, refs.workflow, refs.submission_form_data_pool, refs.submission_copy]
+    joins: [refs.submission_status, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, refs.submission_form_data, refs.step, refs.workflow, refs.daac, refs.submission_form_data_pool, refs.submission_copy]
   },
   where: {
     filters: [
@@ -195,7 +201,7 @@ const getDaacSubmissions = (params) => sql.select({
         fields: fields(allFields),
     from: {
     base: table,
-    joins: [refs.submission_status, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, refs.submission_form_data, refs.step, refs.workflow, refs.submission_form_data_pool, refs.submission_copy]
+      joins: [refs.submission_status, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, refs.submission_form_data, refs.step, refs.workflow, refs.daac, refs.submission_form_data_pool, refs.submission_copy]
     },
         where: {
           filters: [
@@ -224,7 +230,7 @@ const getDaacSubmissions = (params) => sql.select({
                 any: {
                 values: {
                   type: 'select',
-                  fields: ['daac.id'],
+                  fields: ['daac.id', 'daac_name.long_name'],
                   from: {
                   base: 'daac',
                   joins: [{
@@ -254,7 +260,7 @@ const getAdminSubmissions = (params) => sql.select({
   fields: fields(allFields),
   from: {
     base: table,
-    joins: [refs.submission_status, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, refs.submission_form_data, refs.step, refs.workflow, refs.submission_form_data_pool, refs.submission_copy]
+    joins: [refs.submission_status, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, refs.submission_form_data, refs.step, refs.workflow, refs.daac, refs.submission_form_data_pool, refs.submission_copy]
   },
   where: {
     filters: [
@@ -270,16 +276,17 @@ const findAll = ({
   status, created_before, created_after, last_change_before, last_change_after, hidden, sort, order,
   per_page, page, contributor_ids
 }) => sql.select({
-  fields: [fieldMap.id, fieldMap.name, fieldMap.conversation_id, fieldMap.workflow_id, fieldMap.workflow_name, fieldMap.step_name, fieldMap.status, fieldMap.created_at, fieldMap.last_change, fieldMap.hidden, fieldMap.contributor_ids],
+  fields: [fieldMap.id, fieldMap.name, fieldMap.conversation_id, fieldMap.workflow_id, fieldMap.workflow_name, fieldMap.daac_name, fieldMap.step_name, fieldMap.status, fieldMap.created_at, fieldMap.last_change, fieldMap.hidden, fieldMap.contributor_ids],
   from: {
     base: table,
-    joins: [refs.submission_status, refs.step, refs.workflow]
+    joins: [refs.submission_status, refs.step, refs.workflow, refs.daac]
   },
   where: {
     filters: [
       ...(name ? [{ field: 'submission.name', param: 'name' }] : []),
       ...(user_id ? [{ field: 'submission.initiator_edpuser_id', param: 'user_id' }] : []),
       ...(daac_id ? [{ field: 'submission.daac_id', param: 'daac_id' }] : []),
+      ...(daac_name ? [{ field: 'daac.long_name', param: 'daac_name' }] : []),
       ...(workflow_id ? [{ field: 'submission_status.workflow_id', param: 'workflow_id' }] : []),
       ...(workflow_name ? [{ field: 'workflow.long_name', param: 'workflow_name' }] : []),
       ...(step_name ? [{ field: 'step.step_name', param: 'step_name' }] : []),